+++ lib/src/internal/component/base/sbu_base_component.dart
@@ -494,6 +494,8 @@ mixin SBUBaseComponent {
       }
     }
 
+    SendbirdUIKit().customParamsHandler?.onBeforeSendFileMessage(params);
+
     // Use queue manager for sequential file sending
     SBUFileSendQueueManager().addFileTask(
       channel: channel,
@@ -541,6 +543,10 @@ mixin SBUBaseComponent {
           params.replyToChannel = true;
           params.parentMessageId = replyingToMessage.messageId;
         }
+
+        SendbirdUIKit()
+            .customParamsHandler
+            ?.onBeforeSendMultipleFilesMessage(params);
 
         SBUFileSendQueueManager().addMultipleFilesTask(
           channel: channel,